1st commit
This is a simple python script with UI for creating textArt from images.
It allows for creating, copying and saving of textArt.
This tool has basic functions for editing the colours of the image using basic input levels tool.
The top slider is mid-tone Point and the bottom two sliders change white point and black point respectively.
Output is displayed on a text field bellow the sliders.
Size of the font used to display the image can be set by the slider titled font size.